Squash old spi.rs with spi_dma.rs
