+++ src/chips/sam4l/spi.rs
@@ -6,9 +6,16 @@ use hil::spi_master;
 use hil::spi_master::SpiCallback;
 use hil::spi_master::ClockPolarity;
 use hil::spi_master::ClockPhase;
+use dma::DMAChannel;
+use dma::DMAClient;
 
-// Driver for the SPI hardware (seperate from the USARTS), 
-// described in chapter 26 of the SAM4L datasheet)
+/// Implementation of DMA-based SPI master communication for
+/// the Atmel SAM4L CortexM4 microcontroller.
+/// Authors: Sam Crow <samcrow@uw.edu>
+///          Philip Levis <pal@cs.stanfored.edu>
+///
+// Driver for the SPI hardware (seperate from the USARTS),
+// described in chapter 26 of the datasheet
 
 /// The registers used to interface with the hardware
 #[repr(C, packed)]
@@ -46,51 +53,55 @@ pub enum Peripheral {
 }
 
 ///
-/// SPI implementation using the SPI hardware
-/// Supports four peripherals. Each peripheral can have different settings.
-///
-/// The init, read, and write methods act on the currently selected peripheral.
-/// The init method can be safely called more than once to configure different peripherals:
-///
-///     spi.set_active_peripheral(Peripheral::Peripheral0);
-///     spi.init(/* Parameters for peripheral 0 */);
-///     spi.set_active_peripheral(Peripheral::Peripheral1);
-///     spi.init(/* Parameters for peripheral 1 */);
-///
+/// The SAM4L supports four peripherals. 
 pub struct Spi {
-    /// Registers
     regs: *mut SpiRegisters,
-    /// Client
-    callback: Cell<Option<&'static SpiCallback>>,
+    callback: Option<&'static SpiCallback>,
+    dma_read:  Option<&'static mut DMAChannel>,
+    dma_write: Option<&'static mut DMAChannel>,
+    // keep track of which // interrupts are pending in order 
+    // to correctly issue completion event only after both complete.
+    reading: Cell<bool>,
+    writing: Cell<bool>,
 }
 
 pub static mut SPI: Spi = Spi::new();
 
 impl Spi {
-    /// Creates a new SPI object, with peripheral 0 selected
+   /// Creates a new SPI object, with peripheral 0 selected
    pub const fn new() -> Spi {
         Spi {
             regs: SPI_BASE as *mut SpiRegisters,
-            callback: Cell::new(None)
+            callback: None,
+            dma_read:  None,
+            dma_write: None,
+            reading: Cell::new(false),
+            writing: Cell::new(false)
         }
     }
 
     pub fn enable(&self) {
+        self.dma_read.as_ref().map(|read| read.enable());
+        self.dma_write.as_ref().map(|write| write.enable());
         unsafe { volatile_store(&mut (*self.regs).cr, 0b1); }
     }
 
     pub fn disable(&self) {
+        self.dma_read.as_ref().map(|read| read.disable());
+        self.dma_write.as_ref().map(|write| write.disable());
         unsafe { volatile_store(&mut (*self.regs).cr, 0b10); }
     }
 
-    /// Sets the approximate baud rate for the active peripheral
+    /// Sets the approximate baud rate for the active peripheral,
+    /// and return the actual baud rate set.
     ///
-    /// Since the only supported baud rates are 48 MHz / n where n is an integer from 1 to 255,
-    /// the exact baud rate may not be available. In that case, the next lower baud rate will be
-    /// selected.
+    /// Since the only supported baud rates are 48 MHz / n where n 
+    /// is an integer from 1 to 255, the exact baud rate may not 
+    /// be available. In that case, the next lower baud rate will 
+    /// be selected.
     ///
-    /// The lowest available baud rate is 188235 baud. If the requested rate is lower,
-    /// 188235 baud will be selected.
+    /// The lowest available baud rate is 188235 baud. If the 
+    /// requested rate is lower, 188235 baud will be selected.
     pub fn set_baud_rate(&self, rate: u32) -> u32 {
         // Main clock frequency
         let mut real_rate = rate;
@@ -151,8 +162,8 @@ impl Spi {
         }
     }
 
-    /// Returns the value of CSR0, CSR1, CSR2, or CSR3, whichever corresponds to the active
-    /// peripheral
+    /// Returns the value of CSR0, CSR1, CSR2, or CSR3, 
+    /// whichever corresponds to the active peripheral
     fn read_active_csr(&self) -> u32 {
         match self.get_active_peripheral() {
             Peripheral::Peripheral0 => unsafe {volatile_load(&(*self.regs).csr0)},
@@ -161,8 +172,8 @@ impl Spi {
             Peripheral::Peripheral3 => unsafe {volatile_load(&(*self.regs).csr3)},
         }
     }
-    /// Sets the value of CSR0, CSR1, CSR2, or CSR3, whichever corresponds to the active
-    /// peripheral
+    /// Sets the Chip Select Register (CSR) of the active peripheral
+    /// (CSR0, CSR1, CSR2, or CSR3). 
     fn write_active_csr(&self, value: u32) {
         match self.get_active_peripheral() {
             Peripheral::Peripheral0 => unsafe {volatile_store(&mut (*self.regs).csr0, value)},
@@ -171,20 +182,28 @@ impl Spi {
             Peripheral::Peripheral3 => unsafe {volatile_store(&mut (*self.regs).csr3, value)},
         };
     }
+
+    /// Set the DMA channels used for reading and writing.
+    pub fn set_dma(&mut self, read: &'static mut DMAChannel, write: &'static mut DMAChannel) {
+        self.dma_read = Some(read);
+        self.dma_write = Some(write);
+    }
 }
 
 impl spi_master::SpiMaster for Spi {
+    /// By default, initialize SPI to operate at 40KHz, clock is
+    /// idle on low, and sample on the leading edge.
     fn init(&mut self, callback: &'static SpiCallback) {
-        self.callback.set(Some(callback));
-        self.set_rate(40000); // Set initial baud rate to 40kbps
+        self.callback = Some(callback);
+        self.set_rate(40000); // Set initial baud rate to 8MHz
         self.set_clock(ClockPolarity::IdleLow);
         self.set_phase(ClockPhase::SampleLeading);
 
         // Keep slave select active until a last transfer bit is set
-        let mut csr = self.read_active_csr(); 
+        let mut csr = self.read_active_csr();
         csr |= 1 << 3;
         self.write_active_csr(csr);
-
+        
         // Indicate the last transfer to disable slave select 
         unsafe {volatile_store(&mut (*self.regs).cr, 1 << 24)};
         
@@ -192,43 +211,73 @@ impl spi_master::SpiMaster for Spi {
         mode |= 1; // Enable master mode
         mode |= 1 << 4; // Disable mode fault detection (open drain outputs not supported)
         unsafe {volatile_store(&mut (*self.regs).mr, mode)};
+        
     }
 
     fn is_busy(&self) -> bool {
-        false // Since all operations are blocking, never busy
+        self.reading.get() || self.writing.get()
     }
 
+    /// Write a byte to the SPI and return the read; if an
+    /// asynchronous operation is outstanding, do nothing.
     fn read_write_byte(&self, val: u8) -> u8 {
-        self.write_byte(val); 
+        if self.reading.get() || self.writing.get() {
+  //          return 0;
+        }
+        self.write_byte(val);
         // Wait for receive data register full
-        while (unsafe {volatile_load(&(*self.regs).sr)} & 1) == 0 {}
+        while (unsafe {volatile_load(&(*self.regs).sr)} & 1) != 1 {}
         // Return read value
         unsafe {volatile_load(&(*self.regs).rdr) as u8}
     }
-
+      
+    /// Write a byte to the SPI and discard the read; if an
+    /// asynchronous operation is outstanding, do nothing.
     fn write_byte(&self, out_byte: u8) {
+        if self.reading.get() || self.writing.get() {
+ //           return;
+        }
         let tdr = out_byte as u32;
         // Wait for data to leave TDR and enter serializer, so TDR is free
         // for this next byte
         while (unsafe {volatile_load(& (*self.regs).sr)} & 1 << 1) == 0 {}
         unsafe {volatile_store(&mut (*self.regs).tdr, tdr)};
     }
-
+        
+    /// Write 0 to the SPI and return the read; if an
+    /// asynchronous operation is outstanding, do nothing.
     fn read_byte(&self) -> u8 {
         self.read_write_byte(0)
     }
 
-    /// The write buffer has to be mutable because it's passed back to
-    /// the caller, and the caller may want to be able write into it.
+    /// Asynchonous buffer read/write of SPI.
+    /// write_buffer must  be Some; read_buffer may be None;
+    /// if read_buffer is Some, then length of read/write is the
+    /// minimum of two buffer lengths; returns true if operation
+    /// starts (will receive callback through SpiClient), returns
+    /// false if the operation does not start.
+    // The write buffer has to be mutable because it's passed back to
+    // the caller, and the caller may want to be able write into it.
+    //
+    // A better version of this would 
     fn read_write_bytes(&self, 
-                        mut write_buffer:  Option<&'static mut [u8]>, 
-                        mut read_buffer: Option<&'static mut [u8]>,
+                        write_buffer:  Option<&'static mut [u8]>, 
+                        read_buffer: Option<&'static mut [u8]>,
                         len: usize) -> bool {
         let writing = write_buffer.is_some();
         let reading = read_buffer.is_some();
-        if !writing {
+        // If there is no write buffer, or busy, then don't start.
+        // Need to check self.reading as well as self.writing in case
+        // write interrupt comes back first.
+        if !writing  || self.reading.get() || self.writing.get() {
             return false
         }
+
+        // Need to mark if reading or writing so we correctly
+        // regenerate Options on callback
+        self.writing.set(writing);
+        self.reading.set(reading);
+
         let read_len = match read_buffer {
             Some(ref buf) => {buf.len()},
             None          => 0
@@ -239,36 +288,30 @@ impl spi_master::SpiMaster for Spi {
         };
         let buflen = if !reading {write_len}
                      else        {cmp::min(read_len, write_len)};
-        count = cmp::min(buflen, len);
+        let count = cmp::min(buflen, len);
 
-        for i in 0..count {
-            let mut txbyte: u8 = 0;
-            match write_buffer {
-                Some(ref buf) => {txbyte = buf[i];}
-                None          => {}
-            }
-            // Write the value
-            let rxbyte = self.read_write_byte(txbyte);
-            // Don't do a match here due to lifetime problems,
-            // couldn't figure out how to do it -pal
-            if read_buffer.is_some() {
-               let mut buf = read_buffer.as_mut().unwrap();
-               buf[i] = rxbyte;
-            }
+        // The ordering of these operations matters; if you enable then
+        // perform the operation, you can read a byte early on the SPI data register
+        if reading {
+            self.dma_read.as_ref().map(|read| {
+                // We know from the check above that `reading` is only true if
+                // `read_buffer` is `Some`, so `unwrap` is safe here.
+                read.do_xfer_buf(4, read_buffer.unwrap(), count)
+            });
         }
-        self.callback.get().map(|cb| {
-            cb.read_write_done();
-        });
+        self.dma_write.as_ref().map(|write| write.do_xfer_buf(22, write_buffer.unwrap(), count));
+        if reading {
+            self.dma_read.as_ref().map(|read| read.enable());
+        }
+        self.dma_write.as_ref().map(|write| write.enable());
         true
     }
 
-#[allow(unused_variables)]
-    fn set_rate(&self, rate: u32) -> u32 { 
-      self.set_baud_rate(rate)
-    }
-            
-#[allow(unused_variables)]
-    fn set_clock(&self, polarity: ClockPolarity) { 
+    fn set_rate(&self, rate: u32) -> u32 {
+        self.set_baud_rate(rate)
+     }
+
+    fn set_clock(&self, polarity: ClockPolarity) {
         let mut csr = self.read_active_csr();
         match polarity {
             ClockPolarity::IdleHigh => csr |= 1,
@@ -277,8 +320,7 @@ impl spi_master::SpiMaster for Spi {
         self.write_active_csr(csr);
     }
 
-#[allow(unused_variables)]
-    fn set_phase(&self, phase: ClockPhase) { 
+    fn set_phase(&self, phase: ClockPhase) {
         let mut csr = self.read_active_csr();
         match phase {
             ClockPhase::SampleLeading => csr |= 1 << 1,
@@ -287,8 +329,7 @@ impl spi_master::SpiMaster for Spi {
         self.write_active_csr(csr);
     }
 
-    /// Sets the active peripheral
-    fn set_chip_select(&self, cs: u8) -> bool {
+    fn set_chip_select(&self, cs: u8) -> bool{
         if cs >= 4 {
             return false
         }
@@ -311,5 +352,32 @@ impl spi_master::SpiMaster for Spi {
 
     fn clear_chip_select(&self) {
        unsafe {volatile_store(&mut (*self.regs).cr, 1 << 24)};
+    }
+}
+
+impl DMAClient for Spi {
+    fn xfer_done(&mut self, pid: usize) {
+        // I don't know if there are ordering guarantees on the read and
+        // write interrupts, guessing not, so issue the callback when both
+        // reading and writing are complete. In practice it seems like
+        // the read interrupt happens second. -pal
+        // 
+        // The disable calls are commented out because executing them
+        // causes subsequent operations to fail. The call to read_write_bytes
+        // calls enable(), so I don't know why. -pal
+        if pid == 4  { // SPI RX
+           // self.dma_read.as_ref().map(|dma| dma.disable());
+            self.reading.set(false);
+            if !self.reading.get() && !self.writing.get() {
+                self.callback.as_ref().map(|cb| cb.read_write_done());
+            }
+        }
+        if pid == 22 { // SPI TX
+           // self.dma_write.as_ref().map(|dma| dma.disable());
+            self.writing.set(false);
+            if !self.reading.get() && !self.writing.get() {
+                self.callback.as_ref().map(|cb| cb.read_write_done());
+            }
+        }
     }
 }
